adding more automated code refactoring for java 8

+++ core/genometryImpl/src/main/java/com/affymetrix/genometryImpl/ResiduesChars.java
@@ -49,7 +49,7 @@ public enum ResiduesChars {
     public static int getValue(char ch) {
         Integer val = rcMap.get(ch);
         if (val != null) {
-            return val.intValue();
+            return val;
         }
         return -1;
     }
